added modal overlay for users page
