+++ src/container/body/body.container.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import Pagination from "../pagination";
-import Table from "../tableData/table";
+import Table from "../table/tableData/table";
 import styles from "./body.container.module.scss";
 
 const BodyContainer = () => {
