+++ src/container/login/login.module.scss
@@ -8,7 +8,7 @@
   flex-grow: 0;
   display: none;
 
-  @media (min-width: 769px) {
+  @media (min-width: 850px) {
     flex-grow: 1;
     padding: 2em;
     display: flex;
