+++ src/container/table/showTableFilters/filter-table.module.scss
@@ -1,12 +1,53 @@
-.container {
-  width: 270px;
-  height: auto;
-  padding: 20px;
+body.active-modal {
+  overflow-y: auto;
+}
 
-  border-radius: 4px;
-  background: #ffffff;
-  border: 1px solid rgba(84, 95, 125, 0.14);
-  box-shadow: 3px 5px 20px rgba(0, 0, 0, 0.04);
+.buttonModal {
+  display: flex;
+  flex-direction: row;
+  margin-bottom: 18px;
+
+  h4 {
+    color: #545f7d;
+    font-size: 14px;
+    padding-left: 10px;
+  }
+}
+
+.modal,
+.overlay {
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  width: 100vw;
+  height: 100vh;
+  position: fixed;
+}
+
+// .overlay {
+//   background: rgba(49, 49, 49, 0.8);
+// }
+
+.modalContent {
+  line-height: 1.4;
+  max-width: 600px;
+  min-width: 300px;
+  padding: 14px 28px;
+  border-radius: 5px;
+  background-color: white;
+  border: 0.5px solid #545f7d;
+  position: absolute;
+  top: 40%;
+  left: 50%;
+  transform: translate(-140%, -40%);
+}
+
+// body
+.container {
+  padding: 20px;
+  border-radius: 5px;
+  background-color: white;
 }
 
 .organisation {
@@ -26,6 +67,7 @@
   background: #ffffff;
   border: 1px solid #213f7d;
 }
+
 .input {
   height: 40px;
   width: 230px;
@@ -71,32 +113,28 @@
 
 .button {
   display: flex;
+  margin-top: 30px;
   flex-direction: row;
   align-items: center;
-  justify-content: space-between;
   margin-bottom: 30px;
-  margin-top: 30px;
+  justify-content: space-between;
 
   .reset {
     width: 98px;
     height: 40px;
-
-    border: 1px solid #545f7d;
-    border-radius: 8px;
-
     color: #545f7d;
     font-size: 14px;
+    border-radius: 8px;
+    border: 1px solid #545f7d;
   }
 
   .filter {
     width: 98px;
     height: 40px;
-
-    background: #39cdcc;
-    border: 1px solid #39cdcc;
-    border-radius: 8px;
-
     color: #ffffff;
     font-size: 14px;
+    border-radius: 8px;
+    background: #39cdcc;
+    border: 1px solid #39cdcc;
   }
 }
