+++ src/container/table/showUserActions/showuseraction.module.scss
@@ -1,6 +1,21 @@
-.container {
+body.active-modal {
+  overflow-y: hidden;
+}
+
+.btnMdal {
+  padding: 10px 20px;
+  display: block;
+  margin: 100px auto 0;
+  font-size: 18px;
+}
+
+.modal {
   width: 180px;
   height: 130px;
+  position: absolute;
+
+  right: 0;
+
   border-radius: 4px;
   padding-left: 25px;
   padding-top: 25px;
@@ -9,9 +24,15 @@
   box-shadow: 3px 5px 20px rgba(0, 0, 0, 0.04);
 }
 
+.overlay {
+  width: 180px;
+  height: 130px;
+  position: fixed;
+}
+
 .userAction {
   display: flex;
-  margin-bottom: 20px;
+  margin-bottom: 18px;
 
   h4 {
     padding-left: 10px;
