+++ src/layout/topbar/topbar.module.scss
@@ -70,8 +70,8 @@
 // }
 
 .openMenu {
-  width: 35px;
-  height: 35px;
+  width: 45px;
+  height: 45px;
 
   @media (min-width: 600px) {
     display: none;
